Finalised Component 9
This component was ultimately functional, however relied on a seperate instance of running the code. The only changes made to the code was in order to make it more consistent with the test plan, for which the code has had decomposion screenshots of. Work now shifts focus onto the Assembled Outcome.