Assebled Outcome with Components 4 and 9 added with Refinements
This bild of the code only works with expected values, meaning that if a value that is not expecred is used, the code itself crashes. This will hopefully be improved in the fututre, however this build has had decomposition screenshots taken of it and is prepared for more components to be added.